Update  con lo de Harold

+++ 911_RD/911_RD/ARTICULOS.cs
@@ -33,7 +33,7 @@ namespace _911_RD
         public int id_unidad_de_medida { get; set; }
         public string nombre { get; set; }
         public string descripcion { get; set; }
-        public int reorden { get; set; }
+        public double reorden { get; set; }
         public string codigo_barras { get; set; }
         public double precio { get; set; }
         public bool estado { get; set; }
